pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nwo-robotics-api"
version = "0.1.0"
description = "NWO Robotics Layer 5 — Unified API Gateway"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "NWO Robotics", email = "dev@nworobotics.cloud" }]
requires-python = ">=3.11"
dependencies = [
    # API + WebSocket
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",            # Admin dashboard

    # Auth / crypto
    "python-jose[cryptography]>=3.3.0",  # JWT
    "cryptography>=42.0.0",              # ed25519 key ops
    "passlib[bcrypt]>=1.7.4",

    # Database (agent registry + token ledger)
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.1",

    # HTTP proxy (gateway)
    "httpx>=0.27.0",
    "httpx[http2]>=0.27.0",

    # Rate limiting
    "slowapi>=0.1.9",

    # Data / config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "python-ulid>=2.2.0",
    "structlog>=24.2.0",

    # CLI
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.4.9",
]

[project.scripts]
nwo-api = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
